// cgocore-ui/src/app/features/models/loading/views/basic/DatedFlightUldAwbBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldAwbBasicView {

id: number;
selected: boolean;
canDelete: boolean;
datedFlightUldDatedFlightCompanyIataCode: string;
datedFlightUldDatedFlightFlightNumber: string;
datedFlightUldDatedFlightDepartureDate: string;
datedFlightUldUldTypeCode: string;
datedFlightUldUldCode: string;
datedFlightUldUldOwnerCode: string;
airWaybillAirWaybillNumber: string;
numberOfPieces: number;
grossWeight: number;
grossWeightUnitOfMeasureCode: string;
grossVolume: number;
grossVolumeUnitOfMeasureCode: string;

/* Specific Code Start */
/* Specific Code End */
}