// cgocore-ui/src/app/features/models/loading/filters/DatedFlightUldAwbFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldAwbFilter {

datedFlightUldDatedFlightCompanyIataCode: string;
datedFlightUldDatedFlightFlightNumber: string;
datedFlightUldDatedFlightDepartureDateMinValue: string;
datedFlightUldDatedFlightDepartureDateMaxValue: string;
datedFlightUldUldTypeCode: string;
datedFlightUldUldCode: string;
datedFlightUldUldOwnerCode: string;
airWaybillAirWaybillNumber: string;

/* Specific Code Start */
/* Specific Code End */
}