// cgocore-ui/src/app/features/models/loading/sortings/DatedFlightUldAwbSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldAwbSorting {

datedFlightUldDatedFlightCompanyIataCodeOrderType: string;
datedFlightUldDatedFlightFlightNumberOrderType: string;
datedFlightUldDatedFlightDepartureDateOrderType: string;
datedFlightUldUldTypeCodeOrderType: string;
datedFlightUldUldCodeOrderType: string;
datedFlightUldUldOwnerCodeOrderType: string;
airWaybillAirWaybillNumberOrderType: string;
numberOfPiecesOrderType: string;
grossWeightOrderType: string;
grossWeightUnitOfMeasureCodeOrderType: string;
grossVolumeOrderType: string;
grossVolumeUnitOfMeasureCodeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}